Move oteod scripts to ote module
