--- pytorch_toolkit/object_detection/oteod/requirements.txt
@@ -1,1 +1,0 @@
-
