--- pytorch_toolkit/object_detection/oteod/setup.py
@@ -1,7 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(
-    name='oteod',
-    version='0.1',
-    packages=find_packages(),
-)
