+++ pytorch_toolkit/ote/ote/evaluation/common.py
@@ -7,9 +7,8 @@ import tempfile
 
 import yaml
 from mmcv import Config
-from oteod import MMDETECTION_TOOLS
-from oteod.misc import get_complexity_and_size
-from ote.misc import get_file_size_and_sha256
+from ote import MMDETECTION_TOOLS
+from ote.misc import get_file_size_and_sha256 get_complexity_and_size
 
 
 def collect_ap(path):
